# api/storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Promo API",
        "description": "Promo API by PicPay",
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/coupons": {
            "get": {
                "tags": [
                    "coupon"
                ],
                "summary": "Obt\u00e9m todos os cupons",
                "description": "Obt\u00e9m cupons",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Ordena por qualquer atributo da raiz. Ex.: sort=-id, ordena por id DESC",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ids",
                        "in": "query",
                        "description": "Lista de ids separado por v\u00edrgula",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "redirection_type",
                        "in": "query",
                        "description": "Filtra por tipo de redicionamento",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Pesquisa por c\u00f3digo",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "active",
                        "in": "query",
                        "description": "Filtra por active",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "global",
                        "in": "query",
                        "description": "Filtra por global",
                        "required": false,
                        "type": "boolean"
                    },
                    {
                        "name": "campaign_id",
                        "in": "query",
                        "description": "Filtra por id de campanha",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "N\u00famero da p\u00e1gina",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "description": "Tamanho da p\u00e1gina",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "count",
                        "in": "query",
                        "description": "Se retorna o total de itens ou n\u00e3o",
                        "required": false,
                        "type": "boolean"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "coupon"
                ],
                "summary": "Cria cupom",
                "description": "Cria cupom",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/Coupon"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/coupons/{coupon_id}/association": {
            "patch": {
                "tags": [
                    "coupon"
                ],
                "summary": "Associa consumers a um cupom",
                "description": "Associa consumers um cupom",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "coupon_id",
                        "in": "path",
                        "description": "Id do cupom",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/AssociateCampaign"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/coupons/{coupon_id}/status": {
            "patch": {
                "tags": [
                    "coupon"
                ],
                "summary": "Atualiza estado do cupom",
                "description": "Atualiza estado do cupom",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "coupon_id",
                        "in": "path",
                        "description": "Id do coupon",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/UpdateCampaignStatus"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/coupons/{coupon_id}": {
            "put": {
                "tags": [
                    "coupon"
                ],
                "summary": "Atualiza um cupom",
                "description": "Atualiza um cupom",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "coupon_id",
                        "in": "path",
                        "description": "ID do cupom",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "payload",
                        "in": "body",
                        "description": "Os campos 'code' e 'campaign_id' n\u00e3o podem ser editados",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/Coupon"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "AssociateCampaign": {
            "properties": {
                "consumers": {
                    "description": "AssociateCampaignRequest class",
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "example": 3
                    }
                }
            },
            "type": "object"
        },
...
        "Coupon": {
            "properties": {
                "redirection_type": {
                    "description": "CouponRequest class",
                    "type": "string",
                    "enum": [
                        "webview",
                        "app_screen"
                    ]
                },
                "code": {
                    "description": "CouponRequest class",
                    "type": "string",
                    "example": "UBER20"
                },
                "campaign_id": {
                    "description": "CouponRequest class",
                    "type": "string",
                    "example": "5be41adc02655b0018142c7f"
                },
                "global": {
                    "description": "CouponRequest class",
                    "type": "boolean",
                    "example": false
                },
                "max_associations": {
                    "description": "CouponRequest class",
                    "type": "integer",
                    "example": 300
                },
                "webview_url": {
                    "description": "CouponRequest class",
                    "type": "string",
                    "example": "http://cdn.aws.picpay.endereco-grande.com/termos.html"
                },
                "app_screen_path": {
                    "description": "CouponRequest class",
                    "type": "string",
                    "example": "termos"
                },
                "conditions": {
                    "description": "CouponRequest class",
                    "$ref": "#/definitions/ConditionsObject"
                }
            },
            "type": "object"
        },
        "ConditionsObject": {
            "properties": {
                "first_transaction_only": {
                    "description": "CouponRequest class",
                    "type": "boolean",
                    "example": false
                },
                "area_codes": {
                    "description": "CouponRequest class",
                    "type": "array",
                    "items": {
                        "type": "integer",
                        "example": 27
                    }
                }
            },
            "type": "object"
        },
...
        "UpdateCampaignStatus": {
            "properties": {
                "active": {
                    "description": "UpdateCampaignStatusRequest class",
                    "type": "boolean",
                    "example": true
                }
            },
            "type": "object"
        },
...
    }
}